refactor: web server location
