+++ fury/stream/server/__init__.py
@@ -1,3 +1,3 @@
-from fury.stream.server.server import web_server
+from fury.stream.server.main import web_server, web_server_raw_array
 
-__all__ = ['web_server', ]
+__all__ = ['web_server', 'web_server_raw_array']
